+++ .gitignore
@@ -30,7 +30,7 @@ migrate_working_dir/
 .packages
 .pub-cache/
 .pub/
-build/
+/build/
 
 # Symbolication related
 app.*.symbols
