--- lib/extensions/theme.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-
-extension ColorExtension on Color {
-  MaterialColor material() {
-    List strengths = <double>[.05];
-
-    Map<int, Color> swatch = {};
-    final int r = red, g = green, b = blue;
-
-    for (int i = 1; i < 10; i++) {
-      strengths.add(0.1 * i);
-    }
-    for (var strength in strengths) {
-      final double ds = 0.5 - strength;
-      swatch[(strength * 1000).round()] = Color.fromRGBO(
-        r + ((ds < 0 ? r : (255 - r)) * ds).round(),
-        g + ((ds < 0 ? g : (255 - g)) * ds).round(),
-        b + ((ds < 0 ? b : (255 - b)) * ds).round(),
-        1,
-      );
-    }
-    return MaterialColor(value, swatch);
-  }
-}
