--- lib/view/widgets/texture_background.dart
@@ -1,20 +1,0 @@
-import 'package:btech_induction_2023/extensions/system.dart';
-import 'package:flutter/material.dart';
-
-class TextureBackground extends StatelessWidget {
-  const TextureBackground({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: context.height,
-      // width: context.width,
-      decoration: const BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage('images/bg_texture.png'),
-          fit: BoxFit.cover,
-        ),
-      ),
-    );
-  }
-}
